qrcode_gen/requirements.txt
Pillow==9.2.0
qrcode==7.3.1
streamlit==1.11.0
